Reactable bar column

Creates a reactable::colDef object with bar formatting (horizontal bars are added to the background of each cell proprtional to their value scaled by the maximum of the column). Useful for columns of numeric data.

Usage

col_bar(
  df,
  col,
  name = col,
  tooltip = NULL,
  color = "#c5c5c5",
  digits = 0,
  percent = FALSE,
  ...,
  min_width = 75,
  border_right = NULL
)

Arguments

df

dataframe Dataframe of data used in the table.

col

chr Column name of the data (in df).

name

chr Display name of the column to be used as the header. Default is col.

tooltip

chr Tooltip text to be added to the header. Default is NULL.

color

chr CSS appropriate string indicating the bar color.

digits

int Number of decimal digits to display. Default is 0.

percent

bool Whether to format the column as a percent. Default is FALSE.

...

Additional arguments to be passed to reactable::reactable.

min_width

int Minimum width for the column.

border_right

bool / chr Right side border to be added. Default is NULL.