Titles for Echarts grid

Adds titles to echart subplots built using a grid (for example with echarts4r::e_facet(). Note: This function will work with grids specified entirely in percent or entirely in pixels only. It will not work for grids using a mix of pixels and percent. TODO: need to make this work for pixels.

Usage

e_grid_titles(
  e,
  titles,
  font_size = "18px",
  links = NULL,
  link_target = "self",
  top_space = 2
)

Arguments

e

echart Chart to be modified.

titles

vector / list Names to be used. Note: e_grid_titles() will use the order in titles to assign titles along the grid (left to right, top to bottom). So, if you made a facet plot using a charcter grouping variable, you will probably need to sort titles alphabetically for the order to match the grid.

font_size

int / chr Font size to be used. Default is 18px.

links

vector / list Links to be attached to the grid titles. This list follows the same ordering rules as titles. Default is not to add links (NULL).

link_target

string Target to use. Default is 'self'.

top_space

int How far above the plot the title should be placed. The units of top_space (percent vs pixels) will be determined automatically to match the units used in the specification of the grid. Default value is 2.

Examples

groups <- 6
observations <- 10
df <- data.frame(x = rep(1:observations,
                         times = groups),
                 y = runif(groups * observations, 0, 1),
                 group = rep(paste('Group', 1:6), each = observations))

df %>%
  dplyr::group_by(group) %>%
  echarts4r::e_charts(x) %>%
  echarts4r::e_bar(y,
                   name = 'Outcome') %>%
  echarts4r::e_facet(rows = 2,
                     cols = 3,
                     legend_space = 5) %>%
  e_grid_titles(titles = unique(df$group))