Plot leaflet choropleth

Plot interactive choropleth using leaflet with a simplified theme designed for improved appearance in the absence of background tiles. The theme requires pitoolbox to be installed locally. NOTE: if you would like to use the embedded crosstalk functionality of render_map(), you must install a patched version of leaflet from github: devtools::install_github("dmurdoch/leaflet@crosstalk4"). If you do not want to install this version, the crosstalk behavior will silently fail but your code will otherwise run just fine.

Usage

render_map(
  df,
  x,
  map,
  date = NULL,
  id_label = 1,
  info_labels = x,
  palette = NULL,
  collapse_controls = FALSE,
  legend = TRUE,
  legend_position = "bottomright",
  crosstalk = FALSE,
  crosstalk_key = NULL,
  crosstalk_group = NULL,
  ...
)

Arguments

df

dataframe / tibble / datatable Dataset of values for the choropleth.

x

list Named list where keys are the columns of df to plot and the values are the names to be used as the layer and legend titles.

map

sf A simple features dataset containing GIS data to join with df.

date

date Date to be plotted. If NULL, the most recent available date is used. Default is NULL.

id_label

chr Column name in df to use as the shape name in the tooltip.

info_labels

list Named list corresponding to the df columns (keys) and desired label text (values) to appear in the tooltip. By default, the list x will be used.

palette

list List of palettes to be used for the respective variables in x. The list can be named to match the keys of x, if not the palettes will be taken in order. Palettes should be a general palette name (eg: 'Blues'). In the case of factor variables, a vector of specific colors can be given instead. Note: for factor variables, the variable itself must already be ordered in df, otherwise the factors will be ordered by appearance. If no palette is provided, 'Blues' will be used for every layer.

collapse_controls

bool If True and if more than one layer is present (ie, if length(x) > 1), then the layer control in the top right corner will be collapsed. Default is TRUE.

legend

bool If True, the legend will be shown in the bottom right corner. Default is TRUE.

legend_position

chr The desired position of the legend. Default is 'bottomright'.

crosstalk

bool If TRUE, the map data created and used will be converted into a crosstalk SharedData object. Note: enabling crosstalk is mutually exclusive to other aspects of render_map, in particular it cannot be used on maps with multiple layers. Default is FALSE.

crosstalk_key

chr If crosstalk is TRUE, crosstalk_key allows you to optionally specify which column in df to use as the key for the SharedData. Default is NULL, meaning no key will be specified.

crosstalk_group

chr If crosstalk is TRUE, crosstalk_group allows you to specify the group id name provided to SharedData. Default is NULL, meaning a random group name will be generated by crosstalk.

...

Additional arguments to be passed to make_leaflet_labels()

Value

leaflet widget

Examples

if (FALSE) {
library(dplyr)

df_dep %>%
  group_by(dep) %>%
  mutate(p_evo = evolution(p)) %>%
  ungroup() %>%
  render_map(x = c(p = 'Cases',
                   pop = 'Population'),
             id_label = 'dep',
             evolution = c('p'))
}